+++ evals/registry/completion_fns/tmp.yaml
@@ -1,6 +1,6 @@
 # for testing, remove before committing
 
 testing:
-  class: evals.api:OpenAIChatCompletionFn
+  class: evals:OpenAIChatCompletionFn
   args:
     model: gpt-3.5-turbo
